Add On-Premises environmental setup scripts
These scripts enable the installation of Docker and Kubeadm for the apt
package manager.  The Yum package manager is coming RSN.

There is also a Kubernetes deployment script that will deploy Kubernetes
in a fashion that should work for most networks.  Note if your router
network is 192.168.1.0/24, you will need to modify the deployment tools.

Also, please allocate an IP range in the router interface, and configure the
IP range in the metallb-config-073.yaml file.